+++ src/pretalx/orga/context_processors.py
@@ -1,3 +1,4 @@
+import warnings
 from importlib import import_module
 
 from django.conf import settings
@@ -36,9 +37,18 @@ def orga_events(request):
         and hasattr(request, 'user')
         and request.user.is_authenticated
     ):
-        context['nav_event'] = collect_signal(
-            nav_event, {'sender': request.event, 'request': request}
-        )
+        _nav_event = []
+        for receiver, response in nav_event.send_robust(request.event, request=request):
+            if isinstance(response, list):
+                _nav_event += response
+            else:
+                _nav_event.append(response)
+                warnings.warn(
+                    'Please return a list in your nav_event signal receiver, not a dictionary.',
+                    DeprecationWarning
+                )
+
+        context['nav_event'] = _nav_event
 
         if (
             not request.event.is_public